Fix var access settings in BrowserSettings

+++ app/src/main/java/com/hiddenramblings/tagmo/widget/ProgressAlert.kt
@@ -17,15 +17,15 @@ object ProgressAlert {
     private val Number.toPx get() = TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP, this.toFloat(),
         Resources.getSystem().displayMetrics
-    )
+    ).toInt()
 
     private var dialog: AlertDialog? = null
     private var messageText: SoftReference<TextView>? = null
 
     fun show(context: Context, message: String) : ProgressAlert {
         if (dialog == null) {
-            val llvPadding = 10.toPx.toInt()
-            val llhPadding = 8.toPx.toInt()
+            val llvPadding = 10.toPx
+            val llhPadding = 8.toPx
             val ll = LinearLayout(context)
             ll.orientation = LinearLayout.HORIZONTAL
             ll.setPadding(llhPadding, llvPadding, llhPadding, llvPadding)
@@ -38,7 +38,7 @@ object ProgressAlert {
             ll.layoutParams = llParam
             val progressBar = ProgressBar(context)
             progressBar.isIndeterminate = true
-            progressBar.setPadding(0, 0, 12.toPx.toInt(), 0)
+            progressBar.setPadding(0, 0, 12.toPx, 0)
             progressBar.layoutParams = llParam
             llParam = LinearLayout.LayoutParams(
                 ViewGroup.LayoutParams.WRAP_CONTENT,
